create setupIndex helper
