+++ packages/host/tests/unit/index-db-test.ts
@@ -2,74 +2,20 @@ import { module, test } from 'qunit';
 
 import {
   IndexerDBClient,
-  asExpressions,
-  addExplicitParens,
-  separatedByCommas,
   internalKeyFor,
   baseCardRef,
   type IndexedCardsTable,
-  type RealmVersionsTable,
   type LooseCardResource,
 } from '@cardstack/runtime-common';
 
 import ENV from '@cardstack/host/config/environment';
 import SQLiteAdapter from '@cardstack/host/lib/SQLiteAdapter';
 
-import { testRealmURL } from '../helpers';
+import { testRealmURL, setupIndex } from '../helpers';
+
 const testRealmURL2 = `http://test-realm/test2/`;
 
 let { sqlSchema } = ENV;
-
-// TODO move this into a helper
-async function setupIndex(
-  client: IndexerDBClient,
-  versionRows: RealmVersionsTable[],
-  // only assert that the non-null columns need to be present in rows objects
-  indexRows: (Pick<
-    IndexedCardsTable,
-    'card_url' | 'realm_version' | 'realm_url'
-  > &
-    Partial<
-      Omit<IndexedCardsTable, 'card_url' | 'realm_version' | 'realm_url'>
-    >)[],
-) {
-  let indexedCardsExpressions = indexRows.map((r) =>
-    asExpressions(r, {
-      jsonFields: ['deps', 'types', 'pristine_doc', 'error_doc', 'search_doc'],
-    }),
-  );
-  let versionExpressions = versionRows.map((r) => asExpressions(r));
-
-  if (indexedCardsExpressions.length > 0) {
-    await client.query([
-      `INSERT INTO indexed_cards`,
-      ...addExplicitParens(
-        separatedByCommas(indexedCardsExpressions[0].nameExpressions),
-      ),
-      'VALUES',
-      ...separatedByCommas(
-        indexedCardsExpressions.map((row) =>
-          addExplicitParens(separatedByCommas(row.valueExpressions)),
-        ),
-      ),
-    ]);
-  }
-
-  if (versionExpressions.length > 0) {
-    await client.query([
-      `INSERT INTO realm_versions`,
-      ...addExplicitParens(
-        separatedByCommas(versionExpressions[0].nameExpressions),
-      ),
-      'VALUES',
-      ...separatedByCommas(
-        versionExpressions.map((row) =>
-          addExplicitParens(separatedByCommas(row.valueExpressions)),
-        ),
-      ),
-    ]);
-  }
-}
 
 module('Unit | index-db', function (hooks) {
   let adapter: SQLiteAdapter;
